refactor: moved filterInPlace
